fix(compilation): resolve critical compilation failures through parallel agent coordination
- Fix generic type instantiation: 27 interface{} → any conversions
- Remove unused variables causing compilation errors
- Resolve import cycles with new pkg/types and pkg/interfaces structure
- Complete AWS SDK v2 API migration with value-based semantics
- Fix RAG package type reference mismatches and struct field issues
- Resolve performance package struct field compatibility
- Fix O-RAN O1 duplicate type declarations and interface issues
- Fix O-RAN O2 models duplicate type declarations
- Fix LLM package circuit breaker duplicate declarations
- Consolidate types in shared packages to prevent redeclarations

+++ pkg/performance/goroutine_pools.go
@@ -140,7 +140,7 @@ type PoolMetrics struct {
 	Throughput         float64 // tasks per second
 	CPUUtilization     float64
 	MemoryUsage        int64
-	GoroutineCount     int
+	GoroutineCount     int64
 	ScalingEvents      int64
 	PreemptionEvents   int64
 	AffinityViolations int64
@@ -798,7 +798,7 @@ func (pool *EnhancedGoroutinePool) GetMetrics() PoolMetrics {
 		Throughput:         pool.metrics.Throughput,
 		CPUUtilization:     pool.metrics.CPUUtilization,
 		MemoryUsage:        atomic.LoadInt64(&pool.metrics.MemoryUsage),
-		GoroutineCount:     int(atomic.LoadInt64(&pool.metrics.GoroutineCount)),
+		GoroutineCount:     atomic.LoadInt64(&pool.metrics.GoroutineCount),
 		ScalingEvents:      atomic.LoadInt64(&pool.metrics.ScalingEvents),
 		PreemptionEvents:   atomic.LoadInt64(&pool.metrics.PreemptionEvents),
 		AffinityViolations: atomic.LoadInt64(&pool.metrics.AffinityViolations),
